07_pretrained_image_embeddings/pyproject.toml
[tool.poetry]
name = "embedding"
version = "0.1.0"
description = "A tool to compute a variety of image embeddings"
license = "MIT"
authors = ["Justin Swaney <jswaney@mit.edu>"]

[tool.poetry.scripts]
embedding = "embedding:cli"

[tool.poetry.dependencies]
python = "3.10"
click = "^8.1.3"
tensorflow = "^2.9.1"
scikit-image = "^0.19.3"
scikit-learn = "^1.1.2"
umap-learn = "^0.5.3"
seaborn = "^0.11.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
